refactor: group ffi interface in module
