+++ crates/rpc/src/lib.rs
@@ -136,17 +136,37 @@ mod session_wrapper;
 mod handler_traits;
 mod stream_types;
 
-// Process-global runtime for the synchronous convenience wrappers.
-//
-// In native builds SlimRPC owns its own runtime. When the `uniffi` feature is
-// enabled the crate shares the runtime managed by `agntcy-slim-bindings` so that
-// RPC calls run on the same runtime as the owning `App`.
-#[cfg(not(feature = "uniffi"))]
-mod runtime;
-#[cfg(not(feature = "uniffi"))]
-pub use runtime::get_runtime;
+// All `#[uniffi::export]` impls are grouped in this module (compiled only under
+// the `uniffi` feature).
+#[cfg(feature = "uniffi")]
+mod ffi;
+
+// The runtime handle for the synchronous FFI convenience wrappers is only
+// available under the `uniffi` feature (it is owned by `agntcy-slim-bindings`).
+// Native builds have no `get_runtime`: they use the ambient tokio runtime
+// (`Handle::current()`) and the async API directly.
 #[cfg(feature = "uniffi")]
 pub use slim_bindings::get_runtime;
+
+/// Spawn a background task on the runtime that drives streaming RPC work.
+///
+/// Native builds spawn onto the ambient tokio runtime (the caller is always
+/// inside one). The `uniffi` build spawns onto the runtime owned by
+/// `agntcy-slim-bindings`, because FFI callers have no ambient runtime.
+pub(crate) fn spawn<F>(future: F) -> tokio::task::JoinHandle<F::Output>
+where
+    F: std::future::Future + Send + 'static,
+    F::Output: Send + 'static,
+{
+    #[cfg(feature = "uniffi")]
+    {
+        get_runtime().spawn(future)
+    }
+    #[cfg(not(feature = "uniffi"))]
+    {
+        tokio::spawn(future)
+    }
+}
 
 pub use channel::{Channel, MessageContext, MulticastItem};
 pub use codec::{Codec, Decoder, Encoder};
